backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
sqlalchemy==2.0.36
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
slowapi==0.1.9
python-multipart==0.0.12
httpx==0.27.2
websockets==13.1
pywebpush==2.0.0
